+++ package.json
@@ -30,7 +30,7 @@
     "gcs-filepath-validator": "^1.0.0",
     "primus": "^7.1.0",
     "redis": "^2.8.0",
-    "uws": "^8.14.1"
+    "uws": "^9.14.0"
   },
   "devDependencies": {
     "eslint": "^4.9.0",
