+++ src/event-handlers/watch-error.js
@@ -1,0 +1,12 @@
+const displayConnections = require("./display-connections");
+
+module.exports = function(err, filePath, displayId) {
+  console.error(`[${err.message}] [${filePath}] [${displayId}]`);
+
+  displayConnections.sendMessage(displayId, {
+    error: err.code,
+    topic: "watch-result",
+    filePath,
+    msg: `There was an error processing WATCH:${err.message}`
+  });
+}
