+++ test/integration/redis-pubsub.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 const assert = require("assert");
 const rp = require("request-promise-native");
-const gcs = require("../../src/version-compare/gcs");
+const gcs = require("../../src/gcs");
 const simple = require("simple-mock");
 const podname = "test-pod";
 const testPort = 9228;
